pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "Parallel AI agents that fill job applications autonomously"
requires-python = ">=3.11,<4.0"
dependencies = [
    "browser-use>=0.12.0",
    "langchain-anthropic>=0.3.0",
    "anthropic>=0.52.0",
    "pydantic>=2.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "playwright>=1.58.0",
    "openai>=2.16.0",
    "rapidfuzz>=3.14.5",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
